# docs/openapi-payments.yaml
openapi: 3.1.0
info:
  title: Santos Tech — Pagamentos (Pix/Efí)
  version: "1.0.0"
  description: >
    API de pagamentos da Santos Tech (apps/payments-go). Cobra mensalidades
    (recorrência própria) e matrículas/avulsos via Pix pelo gateway Efí
    (efipay). Valores sempre em centavos. Auth: JWT HS256 do ecossistema;
    rotas admin exigem role=3. Erros no padrão `{code,message}`.
servers:
  - url: https://api.santos-tech.com/payments
  - url: http://localhost:3336
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
    Student:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        taxId: { type: string }
        email: { type: string }
        phone: { type: string }
        createdAt: { type: string, format: date-time }
...
    Charge:
      type: object
      properties:
        id: { type: integer }
        kind: { type: string, enum: [mensalidade, matricula, avulso, recorrente] }
        subscriptionId: { type: integer, nullable: true }
        recurrenceId: { type: integer, nullable: true, description: "contrato de PIX Automático (ciclos kind=recorrente)" }
        studentId: { type: integer, nullable: true }
        customerId: { type: integer, nullable: true }
        amountCents: { type: integer }
        dueDate: { type: string, format: date }
        referenceMonth: { type: string, nullable: true }
        status: { type: string, enum: [pending, paid, expired, canceled] }
        provider: { type: string }
        providerChargeId: { type: string }
        correlationId: { type: string }
        publicToken: { type: string, description: "token público p/ tela de pagamento (cobranças de cliente)" }
        brCode: { type: string, description: "Pix copia-e-cola" }
        qrCode: { type: string, description: "imagem do QR (base64/data-uri) ou URL" }
        paidAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
...
paths:
...
  /students:
    post:
      summary: Cria aluno/responsável
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, taxId, email]
              properties:
                name: { type: string }
                taxId: { type: string }
                email: { type: string }
                phone: { type: string }
      responses:
        "201": { description: criado, content: { application/json: { schema: { $ref: "#/components/schemas/Student" } } } }
        "400": { description: inválido, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    get:
      summary: Lista alunos
      responses:
        "200": { description: lista, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Student" } } } } }
  /students/{id}:
    get:
      summary: Detalhe do aluno
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/Student" } } } }
        "404": { description: não encontrado, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /students/{id}/charges:
    get:
      summary: Histórico de cobranças do aluno
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        "200": { description: lista, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Charge" } } } } }
  /plans:
    post:
      summary: Cria plano de mensalidade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, amountCents, dueDay]
              properties:
                name: { type: string }
                amountCents: { type: integer }
                dueDay: { type: integer, minimum: 1, maximum: 28 }
      responses:
        "201": { description: criado, content: { application/json: { schema: { $ref: "#/components/schemas/Plan" } } } }
        "400": { description: inválido, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    get:
      summary: Lista planos
      responses:
        "200": { description: lista, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Plan" } } } } }
  /subscriptions:
    post:
      summary: Cria assinatura (matrícula em plano)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [studentId, planId, amountCents, dueDay]
              properties:
                studentId: { type: integer }
                planId: { type: integer }
                amountCents: { type: integer }
                dueDay: { type: integer, minimum: 1, maximum: 28 }
      responses:
        "201": { description: criada, content: { application/json: { schema: { $ref: "#/components/schemas/Subscription" } } } }
        "400": { description: inválido, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    get:
      summary: Lista assinaturas
      responses:
        "200": { description: lista, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Subscription" } } } } }
  /subscriptions/{id}:
    patch:
      summary: Atualiza status da assinatura
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [active, paused, canceled] }
      responses:
        "200": { description: ok }
        "400": { description: inválido, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /charges:
    post:
      summary: Cria cobrança avulsa/matrícula (Pix)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, studentId, amountCents]
              properties:
                kind: { type: string, enum: [matricula, avulso] }
                studentId: { type: integer }
                amountCents: { type: integer }
                description: { type: string }
                dueDate: { type: string, format: date }
      responses:
        "201": { description: criada (inclui brCode/qrCode), content: { application/json: { schema: { $ref: "#/components/schemas/Charge" } } } }
        "400": { description: inválido, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "502": { description: falha no gateway, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    get:
      summary: Lista cobranças
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [pending, paid, expired, canceled] } }
        - { name: student_id, in: query, schema: { type: integer } }
      responses:
        "200": { description: lista, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Charge" } } } } }
  /charges/{id}:
    get:
      summary: Detalhe da cobrança
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/Charge" } } } }
        "404": { description: não encontrada, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /recurrences:
    post:
      tags: [PIX Automático]
      summary: Cria contrato de PIX Automático de mensalidade (jornada 2)
      description: >
        Cria a recorrência (admin) e gera o QR/copia-e-cola que **só autoriza** o
        débito — nenhuma cobrança na hora; o 1º débito cai no `dueDay` do próximo
        ciclo. Persiste como `pending_auth`, chama `POST /v2/rec` na Efí e devolve o
        `brCode`/`qrCode` de autorização. A transição para `active` chega pelo webhook
        de recorrência após o pagador autorizar.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subscriptionId, amountCents, payerTaxId, payerName, dueDay]
              properties:
                subscriptionId: { type: integer, description: "assinatura vinculada (> 0)" }
                amountCents: { type: integer, description: "valor fixo do ciclo (> 0)" }
                payerTaxId: { type: string, description: "CPF do pagador (11 dígitos)" }
                payerName: { type: string }
                dueDay: { type: integer, minimum: 1, maximum: 28 }
                startDate: { type: string, format: date, description: "default: hoje" }
                endDate: { type: string, format: date, description: "opcional" }
      responses:
        "201":
          description: criada (status, brCode/qrCode de autorização)
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
                  status: { type: string, enum: [pending_auth, active, rejected, expired, canceled] }
                  brCode: { type: string }
                  qrCode: { type: string }
        "400": { description: inválido, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "422": { description: falha no gateway Efí, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    get:
      tags: [PIX Automático]
      summary: Lista contratos de recorrência (admin)
      responses:
        "200": { description: lista, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Recurrence" } } } } }
  /recurrences/{id}:
    get:
      tags: [PIX Automático]
      summary: Detalhe do contrato + ciclos (cobranças vinculadas)
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  recurrence: { $ref: "#/components/schemas/Recurrence" }
                  cycles: { type: array, items: { $ref: "#/components/schemas/Charge" } }
        "400": { description: id inválido, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "404": { description: não encontrada, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /recurrences/{id}/cancel:
    post:
      tags: [PIX Automático]
      summary: Cancela o contrato de recorrência
      description: >
        Cancela na Efí (`PATCH /v2/rec/{idRec}`, best-effort) e marca o contrato como
        `canceled` localmente — o loop de ciclos só cobra contratos `active`, então o
        cancelamento local basta mesmo se a Efí falhar.
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        "200":
          description: cancelado
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, enum: [canceled] }
        "400": { description: id inválido, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "404": { description: não encontrada, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /products:
    post:
      summary: Cria produto do catálogo (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug, name, priceCents]
              properties:
                slug: { type: string }
                name: { type: string }
                description: { type: string }
                priceCents: { type: integer }
      responses:
        "201": { description: criado, content: { application/json: { schema: { $ref: "#/components/schemas/Product" } } } }
        "400": { description: inválido, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    get:
      summary: Lista produtos (admin)
      responses:
        "200": { description: lista, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Product" } } } } }
  /products/{id}:
    put:
      summary: Atualiza produto (admin)
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, priceCents]
              properties:
                name: { type: string }
                description: { type: string }
                priceCents: { type: integer }
                active: { type: boolean }
      responses:
        "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/Product" } } } }
        "400": { description: inválido, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    delete:
      summary: Exclui produto (admin) — histórico de cobranças preservado (product_id vira null)
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        "200": { description: excluído, content: { application/json: { schema: { type: object, properties: { ok: { type: boolean } } } } } }
        "404": { description: não encontrado, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /products/by-slug/{slug}:
    get:
      summary: Produto público por slug (sem auth; só ativos)
      security: []
      parameters: [{ name: slug, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/Product" } } } }
        "404": { description: não encontrado, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /me/customer:
    get:
      summary: Customer do usuário logado (cria se não existir)
      responses:
        "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/Customer" } } } }
    put:
      summary: Atualiza CPF/telefone do customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                taxId: { type: string }
                phone: { type: string }
      responses:
        "200": { description: ok }
  /me/cart:
    get:
      summary: Carrinho do usuário (itens enriquecidos com o produto)
      responses:
        "200":
          description: linhas do carrinho
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    product: { $ref: "#/components/schemas/Product" }
                    quantity: { type: integer }
    post:
      summary: Adiciona produto ao carrinho (por slug)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug]
              properties:
                slug: { type: string }
      responses:
        "200": { description: ok }
        "404": { description: produto não encontrado, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /me/cart/{productId}:
    delete:
      summary: Remove produto do carrinho
      parameters: [{ name: productId, in: path, required: true, schema: { type: integer } }]
      responses:
        "200": { description: ok }
  /me/cart/checkout:
    post:
      summary: Fecha o carrinho e gera a cobrança Pix
      description: >
        Recalcula o total a partir do catálogo (preço sempre do servidor),
        cria a cobrança Efí, persiste os itens, esvazia o carrinho e
        devolve o token público + brCode/qrCode.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [taxId]
              properties:
                taxId: { type: string, description: "CPF do pagador" }
                phone: { type: string }
                save: { type: boolean, description: "salva CPF/telefone no customer" }
      responses:
        "201":
          description: cobrança criada
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  brCode: { type: string }
                  qrCode: { type: string }
                  amountCents: { type: integer }
        "400": { description: carrinho vazio ou CPF ausente, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "502": { description: falha no gateway, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /me/subscribe:
    post:
      summary: Assina um produto recorrente (checkout de assinatura via PIX Automático)
      description: >
        Checkout dedicado a um produto recorrente (item único, fora do carrinho).
        Valida que o produto existe e tem recurring=true, faz upsert do customer por
        (conta, CPF) e cria, via PIX Automático Jornada 3, a recorrência (status
        pending_auth) MAIS a 1ª cobrança com vencimento hoje — devolvendo UM ÚNICO
        QR/copia-e-cola que autoriza a recorrência e paga a 1ª parcela. O pagador é
        levado a /pay/{token}; o webhook de recorrência move o contrato para active ao
        autorizar e o webhook pix marca a 1ª cobrança como paid. Mesma forma de resposta
        do /me/cart/checkout.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId, taxId, name, email]
              properties:
                productId: { type: integer, description: "ID do produto recorrente" }
                taxId: { type: string, description: "CPF do pagador (11 dígitos)" }
                phone: { type: string, description: "telefone (10–11 dígitos)" }
                name: { type: string }
                email: { type: string, format: email }
                save: { type: boolean, description: "salva dados no frontend (localStorage)" }
      responses:
        "201":
          description: assinatura criada (recorrência + 1ª cobrança)
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string, description: "token público da 1ª cobrança (tela /pay/{token})" }
                  brCode: { type: string, description: "copia-e-cola único de autorização+pagamento" }
                  qrCode: { type: string, description: "imagem do QR (base64/data-uri)" }
                  amountCents: { type: integer, description: "valor da 1ª parcela (= preço do produto)" }
        "400": { description: validação (productId/CPF/email/telefone/nome), content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "404": { description: produto não encontrado, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "409": { description: "produto não é recorrente (not_recurring) ou mal configurado (invalid_product)", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "422": { description: falha no gateway Efí, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
  /pay/{token}:
    get:
      summary: Dados públicos da cobrança (tela de pagamento)
      security: []
      parameters: [{ name: token, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  amountCents: { type: integer }
                  brCode: { type: string }
                  qrCode: { type: string }
                  status: { type: string, enum: [pending, paid, expired, canceled] }
                  dueDate: { type: string, format: date }
        "404": { description: não encontrada, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /pay/{token}/events:
    get:
      summary: SSE do status da cobrança (Server-Sent Events)
      security: []
      description: >
        text/event-stream. Emite `event: status` com o status atual no início;
        quando o webhook confirmar o pagamento, emite `event: paid` e encerra.
        Usa Redis pub/sub no canal `pay:charge:{token}`.
      parameters: [{ name: token, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: stream SSE (text/event-stream) }
        "404": { description: não encontrada, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /analytics:
    get:
      summary: Dashboard financeiro — KPIs + séries temporais + distribuições (admin)
      description: >
        Agrega cobranças de `pay_charges` no período `range` e devolve KPIs,
        timeseries por dia/mês, distribuição por status e por tipo, e top produtos.
        Campos monetários sempre em centavos. `deltaPaidPct` compara o período atual
        com o anterior de mesmo tamanho.
      parameters:
        - name: range
          in: query
          required: false
          schema:
            type: string
            enum: [7d, 30d, 90d, 12m, all]
            default: 30d
          description: Janela de tempo do dashboard.
      responses:
        "200":
          description: analytics calculados
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Analytics" }
        "500": { description: falha no banco, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /charges/{id}/receipt:
    get:
      summary: Comprovante Pix da cobrança (PDF)
      tags: [Efí]
      description: >
        Baixa o comprovante Pix da cobrança no formato PDF (bytes crus), consultado
        diretamente em `GET /v2/gn/pix/comprovantes?txid=<correlationId>` da API Efí.
        Requer a cobrança no status `paid`; status diferente retorna 409 `not_paid`.
        Requer role=3 (Admin).
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer }, description: ID da cobrança }
      responses:
        "200":
          description: Comprovante em PDF
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "401": { description: não autenticado, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "403": { description: acesso proibido (não é admin), content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "404": { description: cobrança não encontrada, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "409":
          description: cobrança ainda não paga (not_paid)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "502": { description: falha ao obter comprovante na Efí, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /efi/balance:
    get:
      summary: Saldo disponível na conta Efí
      tags: [Efí]
      description: >
        Retorna o saldo disponível na conta Efí/Pix em centavos, consultado
        diretamente em `GET /v2/gn/saldo` da API Efí. Requer role=3 (Admin).
      responses:
        "200":
          description: Saldo retornado com sucesso
          content:
            application/json:
              schema:
                type: object
                required: [availableCents]
                properties:
                  availableCents:
                    type: integer
                    format: int64
                    description: Saldo em centavos (ex. 10000 = R$ 100,00)
                    example: 10000
        "401": { description: não autenticado, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "403": { description: acesso proibido (não é admin), content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "502": { description: falha ao consultar a Efí, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
  /efi/reports:
    post:
      summary: Solicitar relatório de extrato de conciliação Pix
      tags: [Efí]
      description: >
        Solicita a geração de um relatório de extrato de conciliação Pix via
        `POST /v2/gn/relatorios/extrato-conciliacao`. A Efí processa de forma assíncrona
        e retorna um `reportId`. Use `GET /efi/reports/{id}` para consultar o status e
        baixar o CSV quando pronto. Schema confirmado: `tipoRegistros.pixRecebido=true`.
        Endpoint desabilitado em homologação (500 "Funcionalidade desabilitada") — ativo em produção.
        Requer role=3 (Admin).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [date]
              properties:
                date:
                  type: string
                  format: date
                  example: "2025-06-18"
                  description: Data do movimento no formato YYYY-MM-DD
      responses:
        "200":
          description: ID do relatório solicitado (a Efí processa de forma assíncrona)
          content:
            application/json:
              schema:
                type: object
                required: [reportId]
                properties:
                  reportId:
                    type: string
                    description: ID do relatório para consultar via GET /efi/reports/{id}
        "400": { description: data inválida ou ausente, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "401": { description: não autenticado, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "403": { description: acesso proibido (não é admin), content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "502": { description: falha ao consultar a Efí, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...